requirements.txt
streamlit>=1.30.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0

# AI Providers (installer seulement ceux que tu utilises)
openai>=1.0.0
anthropic>=0.18.0

# Optional
langchain>=0.1.0
langchain-community>=0.0.20
chromadb>=0.4.22

pandas>=2.0.0
plotly>=5.18.0